BM_VM/Platforms/VAG/udacity/requirements.txt
findspark
#pyspark
python-dateutil
pathlib
kafka-python
